requirements.txt
# Core ML Libraries (Updated versions)
tensorflow==2.19.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1
joblib==1.4.2

# API Development
flask==3.0.3
flask-cors==5.0.0
gunicorn==23.0.0
requests==2.32.3

# Data Processing
python-dotenv==1.0.1
PyYAML==6.0.2
openpyxl==3.1.5

# Development & Testing
pytest==8.3.3
black==24.10.0
jupyter==1.1.1
